backend/requirements.txt
# Django Core
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
django-extensions==3.2.3

# Database
psycopg[binary]==3.1.18
psycopg[pool]==3.1.18
django-postgres-extra==2.0.9
postgis==1.0.4

# Cache & Queue
redis==5.0.4
django-redis==5.4.0
rq==1.15.1
django-rq==2.9.0

# Message Queue
kafka-python==2.0.2

# Object Storage
minio==7.2.7
django-minio-backend==2.0.0

# Video Processing
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.12.0
Pillow==10.2.0

# ML & Analytics
ultralytics==8.2.29
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.4.2
pandas==2.2.2

# Web Framework
gunicorn==21.2.0
uvicorn==0.29.0
daphne==4.1.2

# Environment & Utilities
python-dotenv==1.0.1
celery==5.4.0
flower==2.0.1
requests==2.31.0
httpx==0.27.0

# Monitoring
prometheus-client==0.20.0
django-prometheus==2.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.4.0

# Serialization
PyYAML==6.0.1
python-magic==0.4.27

# Logging
structlog==24.4.0
sentry-sdk==2.2.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
Faker==24.3.0

# Code Quality
black==24.4.2
flake8==7.0.0
isort==5.13.2